collect creation logic of ui parameter to UseCase
