+++ app/src/main/java/com/example/droidkaigi/conf2020app/UseCase.kt
@@ -1,4 +1,3 @@
 package com.example.droidkaigi.conf2020app
 
-
 object UseCase
