+++ app/src/main/java/com/example/droidkaigi/conf2020app/ui/UiSession.kt
@@ -2,11 +2,8 @@ package com.example.droidkaigi.conf2020app.ui
 
 import androidx.ui.graphics.Color
 import com.example.droidkaigi.conf2020app.data.response.*
-import java.text.SimpleDateFormat
-import java.time.LocalDate
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
-import java.util.*
 
 data class UiSession(
     val id: SessionId,
@@ -19,6 +16,7 @@ data class UiSession(
 
     val endsAt: LocalDateTime,
     val startsAt: LocalDateTime,
+    val dateFromTo: String,
     val lengthInMinutes: Int,
 
     val room: Room,
@@ -30,14 +28,23 @@ data class UiSession(
     val interpretationTarget: Boolean,
     val isPlenumSession: Boolean,
     val isServiceSession: Boolean,
-    val asset: Asset
-)
+    val asset: Asset,
+
+    val listItemParam: ListItemParam
+) {
+    data class ListItemParam(
+        val widthRate: Float
+    )
+}
+
+data class SessionId(val value: String)
+
 val UiSession.titleText: String
     get() = title.ja
 val UiSession.roomNameText: String
     get() = room.name.ja
 
-
+fun LocalDateTime.toFString(formatter: DateTimeFormatter): String = this.format(formatter)
 
 fun UiSession.toPrintString() = """
 # ${room.name.ja} ($lengthInMinutes min)
@@ -53,59 +60,3 @@ message: $message
             
         """.trimIndent()
 
-data class SessionId(val value: String)
-
-
-fun TimeTable.toUiSessions(): List<UiSession> = let { timeTable ->
-    timeTable.sessions.map { session ->
-        val room = timeTable.rooms.first { it.id == session.roomId }
-        val category = timeTable.categories.firstOrNull { it.id == session.sessionCategoryItemId }
-        val speakers = timeTable.speakers.filter { speaker -> session.speakers.any { id -> id == speaker.id } }
-
-        UiSession(
-            id = SessionId(session.id),
-            title = session.title,
-            language = session.language,
-            description = session.description,
-            speakers = speakers,
-            message = session.message?.ja,
-            targetAudience = session.targetAudience,
-
-            endsAt = session.endsAt.toLocaleDateTime(),
-            startsAt = session.startsAt.toLocaleDateTime(),
-            lengthInMinutes = session.lengthInMinutes,
-
-            room = room,
-            roomColor = when(room.name.ja) {
-                "App bars" -> RoomColors.appBars
-                "Backdrop" -> RoomColors.backDrop
-                "Cards" -> RoomColors.cards
-                "Dialogs" -> RoomColors.dialogs
-                "Exhibition" -> RoomColors.exhibition
-                "Pickers" -> RoomColors.pickers
-                "Sliders" -> RoomColors.slides
-                "Tabs" -> RoomColors.tabs
-                else -> RoomColors.empty
-            },
-            sessionCategoryItem = category,
-            sessionType = session.sessionType,
-
-            levels = session.levels,
-            interpretationTarget = session.interpretationTarget,
-            isPlenumSession = session.isPlenumSession,
-            isServiceSession = session.isServiceSession,
-            asset = session.asset
-        )
-    }
-}
-
-
-private val dateFormat =
-    SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ssXXX", Locale.JAPAN)
-
-private val localDateFormat =
-    DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ssXXX")
-
-private fun String.toDate() = dateFormat.parse(this)
-private fun String.toLocaleDate() = LocalDate.parse(this, localDateFormat)
-private fun String.toLocaleDateTime() = LocalDateTime.parse(this, localDateFormat)
